+++ core/gpslocation.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "core/gpslocation.h"
 #include "core/divesite.h"
+#include "core/dive.h"
 #include "qt-models/gpslistmodel.h"
 #include "core/pref.h"
 #include "core/qthelper.h"
