+++ core/save-git.c
@@ -17,6 +17,7 @@
 #include <fcntl.h>
 #include <git2.h>
 
+#include "dive.h"
 #include "divesite.h"
 #include "subsurface-string.h"
 #include "trip.h"
