+++ tests/testgitstorage.cpp
@@ -2,6 +2,7 @@
 #include "testgitstorage.h"
 #include "git2.h"
 
+#include "core/dive.h"
 #include "core/divesite.h"
 #include "core/file.h"
 #include "core/qthelper.h"
