+++ tests/testrenumber.cpp
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "testrenumber.h"
+#include "core/dive.h"
 #include "core/divesite.h"
 #include "core/trip.h"
 #include "core/file.h"
